% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_color.default.R
\name{change_color.default}
\alias{change_color.default}
\title{Does not change the color of the trace plot if default color is used.}
\usage{
\method{change_color}{default}(template, color)
}
\arguments{
\item{template}{A list from the observed_data.R function created using trace_presets variable to identify the trace plot getting added.}
}
\value{
Returns a stop message
}
\description{
Does not change the color of the trace plot if default color is used.
}
