% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_color.observed_data.R
\name{change_color.observed_data}
\alias{change_color.observed_data}
\title{Accepts a list containing a template for a trace of a bar plot for the observed data plot and set the color of the plot.}
\usage{
\method{change_color}{observed_data}(template, color)
}
\arguments{
\item{template}{A list from the observed_data.R function created using trace_presets variable to identify the trace plot getting added.}

\item{color}{A hex value to set as the color.}
}
\value{
Sets color of the trace plot template added as argument to the function
}
\description{
Accepts a list containing a template for a trace of a bar plot for the observed data plot and set the color of the plot.
}
