% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_color.R
\name{change_color}
\alias{change_color}
\title{Accepts a list containing a template for a trace of a plot and set the color of the plot using the method change_color.}
\usage{
change_color(template, ...)
}
\arguments{
\item{template}{A list from the observed_data.R function created using trace_presets variable to identify the trace plot getting added.}
}
\value{
Sets color of the trace plot template added as argument to the function
}
\description{
Accepts a list containing a template for a trace of a plot and set the color of the plot using the method change_color.
}
