% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_color.var_data.R
\name{change_color.var_data}
\alias{change_color.var_data}
\title{Accepts a list containing a template for a trace of a scatter/line plot for the trace plot being added and set the color of the plot.}
\usage{
\method{change_color}{var_data}(template, color, width)
}
\arguments{
\item{template}{A list from the observed_data.R function created using trace_presets variable to identify the trace plot getting added.}

\item{color}{A hex value to set as the color.}

\item{width}{A numeric value that indicates the width of the plot.}
}
\value{
Sets color of the trace plot template added as argument to the function
}
\description{
Accepts a list containing a template for a trace of a scatter/line plot for the trace plot being added and set the color of the plot.
}
