% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_historic_forecast_dataset.R
\name{create_historic_forecast_dataset}
\alias{create_historic_forecast_dataset}
\title{Accepts incoming projection with observed data and extracts 7th and 14th day of median projections along with confidence intervals.}
\usage{
create_historic_forecast_dataset(
  data,
  days = c(7, 14),
  projection_variable = "median",
  ci = 90,
  obs_column
)
}
\arguments{
\item{data}{A data frame specif.}

\item{days}{A vetor specifying which days to extract values for.}

\item{projection_variable}{A character string specifies which projection variabe to extract. Example: "median" or "mean".}

\item{ci}{A numeric value specifying the confidence interval.}

\item{obs_column}{A character string specifying the name of column for observed variable.}
}
\value{
Returns a data frame containing the EpiNow projections and observed values for only the specified days in the days argument.
}
\description{
Accepts incoming projection with observed data and extracts 7th and 14th day of median projections along with confidence intervals.
}
