% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_creation.R
\name{dataset_creation}
\alias{dataset_creation}
\title{Creates a list of R objects to be used for EpiNow plotting using plotly function in observed_data.R script.}
\usage{
dataset_creation(
  dataset =
    "C:/Users/saran/OneDrive/Documents/WasteWaterCollabAnalysis/wastewater-collaborative-analysis/Data/Observed_data/Ottawa_WWTP_Extended_Aggregated_Data_MR.csv",
  site = "Ottawa WWTP",
  enddate = "2022-02-01",
  y_col = "nmN1N2_7dma",
  column = c("sampleDate", "siteID", "nmN1N2_7dma")
)
}
\arguments{
\item{dataset}{A data frame.}

\item{site}{A character string specifying the name of waste water site to be used to create plot.}

\item{enddate}{A date value specifying the value of end date for Ottawa Waste water data.}

\item{y_col}{A character string specifying the name of column from observed data to plot.}

\item{column}{A vector that contains the name of columns to filter from the input data and passed as argument to data_prep function}
}
\value{
The list of R objects to be used in the plotting function in observed_data.R.
\itemize{
  \item cleaned_data - The input dataset that has been cleaned by selecting the right siteID and columns and right formatting of date.
  \item cleaned_march_data - Based on site = "Ottawa WWTP", The input dataset has been cleaned by selecting the right siteID and columns and right formatting of date and filtering data from month of March.
  \item waste_roc - A data frame that has change and percent change in signal from previous day along with its log values.
  \item y_col - A character value specifying the wastewater signal column being used to plot in the dataset.
  \item "daily_viral_signal_call" - A list that is used to provide arguments to plotting function in observed_data.R for the wastewater signal column.
  \item "daily_roc_call_log_mag" - A list that is used to provide arguments to plotting function in observed_data.R for the column that contains the log of daily change in wastewater signal.
  \item "daily_roc_call_log_pct" - A list that is used to provide arguments to plotting function in observed_data.R for the column that contains the log of daily % change in wastewater signal.
  \item "daily_roc_call_pct" - A list that is used to provide arguments to plotting function in observed_data.R for the column that contains the daily % change in wastewater signal.
}
}
\description{
Creates a list of R objects to be used for EpiNow plotting using plotly function in observed_data.R script.
}
\examples{
dataset_creation(dataset = "C:/Users/saran/OneDrive/Documents/WasteWaterCollabAnalysis/wastewater-collaborative-analysis/Data/Observed_data/Toronto_WWTP_Extended_Aggregated_Data_MR.csv",site = "NT", y_col = "nmN1N2_7dma")
}
