% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_for_sites.R
\name{forecast_for_sites}
\alias{forecast_for_sites}
\title{Accepts incoming data, cleans it based on the site provided and calls short_term_forecast function to run EpiNow projections.}
\usage{
forecast_for_sites(
  dataset = "Data/Observed_data/OPH_Observed_COVID_Data.csv",
  site = "Ottawa WWTP",
  startdate = as.Date("2021-02-01"),
  horizon = 14,
  enddate = as.Date("2022-07-16"),
  projecting = "cases",
  column = c("sampleDate", "siteID", "nmN1N2_7dma"),
  input_multiplier = 1,
  project_hospital = FALSE,
  y_col = "N1_N2_avg_clean"
)
}
\arguments{
\item{dataset}{A data frame.}

\item{site}{A character string specifying the site for waste water or Covid case data OR the column in the dataset in case of absenteeism data.}

\item{startdate}{A date value specifying the start date for training data for EpiNow projections}

\item{horizon}{A numeric value specifying the number of days to forecast into future.}

\item{enddate}{A date value specifying the end date for training data for EpiNow projections.}

\item{projecting}{A character string specifies whether we are projecting wastewater or cases. Accepted values: "ww" or "cases"}

\item{column}{A vector that provides the list of variables to be selected from waste water data.}

\item{input_multiplier}{A numeric value which gets multiplied to the input variable that is being projected.}

\item{project_hospital}{A boolean value that determines whether projections are being made on hospital data or Covid case data.}

\item{y_col}{A character string specifying the variable to perform projections on from the wastewater data.}
}
\value{
Saves the forecast data frame or list of summary estimates or Both into Data/Projections directory. Data directory is in root folder.
}
\description{
Accepts incoming data, cleans it based on the site provided and calls short_term_forecast function to run EpiNow projections.
}
