% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals_projections.R
\name{intervals_projections}
\alias{intervals_projections}
\title{Accepts incoming data, cleans it, run two week rolling forecast using EpiNow where the initial start date is incremented in each interval to cover approximately 20, 12, 9, 6, 3, and 1 month of training data in each interval.}
\usage{
intervals_projections(
  dataset =
    "https://raw.githubusercontent.com/Big-Life-Lab/PHESD/main/Wastewater/Ottawa/Data/wastewater_virus.csv",
  site = "Ottawa WWTP",
  y_col = "N1_N2_avg_clean",
  startdate = as.Date("2020-07-01"),
  enddate = as.Date("2022-02-01"),
  column = c("sampleDate", "siteID", "nmN1N2_7dma"),
  date_intervals = c(0, 137, 229, 321, 419, 470),
  input_multiplier = 1e+06,
  projecting = "ww"
)
}
\arguments{
\item{dataset}{A data frame.}

\item{site}{A character string specifying the site for waste water or Covid case data OR the column in the dataset in case of absenteeism data.}

\item{y_col}{A character string specifying the variable to perform projections on from the wastewater data.}

\item{startdate}{A date value specifying the initial start date for training data for validation using EpiNow.}

\item{enddate}{A date value specifying the end date for training data for rolling predictions using EpiNow.}

\item{column}{A vector that provides the list of variables to be selected from waste water data. Date column must be named "sampleDate" in the input dataset to work with this function.}

\item{date_intervals}{A vector that specifies the numeric values to add to the start date of the data in each interval in the for loop.}

\item{input_multiplier}{A numeric value which gets multiplied to the input variable that is being projected.}

\item{projecting}{A character string specifies whether we are projecting waste water or cases. Accepted values: "ww" or "cases"}
}
\value{
Saves 6 different forecast data frame into Data/Projections directory. Data directory is in root folder. The data saved contains entire forecast, projections filtered by growth rate and reported_cases, and filtered by 7th and 14th day of projections for growth rate and reported cases in separate data frames.
}
\description{
Accepts incoming data, cleans it, run two week rolling forecast using EpiNow where the initial start date is incremented in each interval to cover approximately 20, 12, 9, 6, 3, and 1 month of training data in each interval.
}
