% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_fig1.R
\name{normalization_fig1}
\alias{normalization_fig1}
\title{Accepts incoming data and creates plotly plots to show the missingness of data in the variables for mN1, mN2, and mE genes viral signal in the Ontario data for different sites.}
\usage{
normalization_fig1(
  data = c("../../Data/Observed_data/Waterloo_WWTP_Extended_Aggregated_Data_MR.csv",
    "../../Data/Observed_data/Ottawa_WWTP_Extended_Aggregated_Data_rev_MR.csv",
    "../../Data/Observed_data/Ryerson_WWTP_Extended_Aggregated_Data_VP.csv",
    "../../Data/Observed_data/Toronto_WWTP_Extended_Aggregated_Data_MR.csv",
    "../../Data/Observed_data/Windsor_WWTP_Extended_Aggregated_Data_MR.csv"),
  arrow_x_end = c(as.Date("2020-07-14"), as.Date("2020-10-12"), as.Date("2021-01-10"),
    as.Date("2021-04-10"), as.Date("2021-07-09"), as.Date("2021-10-07"),
    as.Date("2022-01-05"), as.Date("2022-04-05"), as.Date("2022-07-04"),
    as.Date("2022-09-22")),
  text_display = c("2020-07-14", "2020-10-12", "2021-01-10", "2021-04-10", "2021-07-09",
    "2021-10-07", "2022-01-05", "2022-04-05", "2022-07-04", "2022-09-22"),
  figwidth = 1500,
  figheight = 8500,
  margin_subplot = 0.005,
  n1_var_name = "mN1",
  n2_var_name = "mN2",
  sitedesc_var = "sys_siteDesc",
  date_var_name = "sampleDate"
)
}
\arguments{
\item{data}{A vector with path of wastewater datasets to be used from different labs.}

\item{arrow_x_end}{A vector of date values to specify where to add an arrow to show the date breaks on the plot.}

\item{text_display}{A vector the same dates as in arrow_x_end vector in form of character values to add as text to the plot.}

\item{figwidth}{A numeric value specifying the width of the plot.}

\item{figheight}{A numeric value specifying the height of the plot.}

\item{margin_subplot}{A numeric value specifying the distance between each subplot for different sites.}
}
\value{
A list containing plotly plot showing missingness of data in the variables for the variables mN1, mN2, and mE viral signal in wastewater for different sites in subplots and the combined data with all labs.
}
\description{
Accepts incoming data and creates plotly plots to show the missingness of data in the variables for mN1, mN2, and mE genes viral signal in the Ontario data for different sites.
}
