% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_ww_data.R
\name{normalized_ww_data}
\alias{normalized_ww_data}
\title{Accepts incoming wastewater data and normalizes the value of the wastewater signal.}
\usage{
normalized_ww_data(dataset, site, y_column, start_date, end_date)
}
\arguments{
\item{dataset}{A data frame.}

\item{site}{A character string specifying the site ID to filter}

\item{y_column}{A character string specifying the name of the wastewater signal column to normalize.}

\item{start_date}{A date value specifying the start date in the dataset to filter.}

\item{end_date}{A date value specifying the end date in the dataset to filter.}
}
\value{
The final normalized dataframe where the wastewater signal column is multiplied with 1000000.
}
\description{
Accepts incoming wastewater data and normalizes the value of the wastewater signal.
}
