% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections_validation.R
\name{projections_validation}
\alias{projections_validation}
\title{Accepts incoming data, cleans it, run two week rolling forecast using EpiNow where the initial start and end date is incremented by 7 days at every step. 7th and 14th day of forecast values from different rolling periods have been extracted for analysis.}
\usage{
projections_validation(
  dataset =
    "https://raw.githubusercontent.com/Big-Life-Lab/PHESD/main/Wastewater/Ottawa/Data/wastewater_virus.csv",
  site = "Ottawa WWTP",
  startdate = as.Date("2020-07-01"),
  enddate = as.Date("2020-11-01"),
  y_col = "N1_N2_avg_clean",
  project_until = as.Date("2022-07-10"),
  projecting = "cases",
  input_multiplier = 1,
  column = c("sampleDate", "siteID", "nmN1N2_7dma")
)
}
\arguments{
\item{dataset}{A data frame.}

\item{site}{A character string specifying the site for waste water or Covid case data OR the column in the dataset in case of absenteeism data.}

\item{startdate}{A date value specifying the initial start date for training data for validation using EpiNow.}

\item{enddate}{A date value specifying the initial end date for training data for validation using EpiNow.}

\item{y_col}{A character string specifying the variable to perform projections on from the wastewater data.}

\item{project_until}{A date value specifying the maximum value for final end date for the rolling projections.}

\item{projecting}{A character string specifies whether we are projecting waste water or cases. Accepted values: "ww" or "cases"}

\item{input_multiplier}{A numeric value which gets multiplied to the input variable that is being projected.}

\item{column}{A vector that provides the list of variables to be selected from waste water data}
}
\value{
Saves 6 different forecast data frame into Data/Projections directory. Data directory is in root folder. The data saved contains entire forecast, projections filtered by growth rate and reported_cases, and filtered by 7th and 14th day of projections for growth rate and reported cases in separate data frames.
}
\description{
Accepts incoming data, cleans it, run two week rolling forecast using EpiNow where the initial start and end date is incremented by 7 days at every step. 7th and 14th day of forecast values from different rolling periods have been extracted for analysis.
}
