% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reworked_figure.R
\name{reworked_figure}
\alias{reworked_figure}
\title{Accepts incoming data and creates plotly plots to show trend in waste water data and compare observed and predicted plots.}
\usage{
reworked_figure(
  xaxis,
  yaxis,
  yaxis2 = NULL,
  yaxis_button = FALSE,
  yaxis2_button = FALSE,
  y_button_name = "",
  y2_button_name = "",
  error_bands = FALSE,
  error_data = NULL,
  error_pct = FALSE,
  error_col = NULL,
  error_bars = FALSE,
  level = FALSE,
  level_upper = NULL,
  level_lower = NULL,
  level_col = NULL,
  titles,
  vline = FALSE,
  vline_date = NULL,
  smooth = FALSE,
  smooth_y = NULL,
  smooth_bandwidth = NULL,
  smooth_name = NULL,
  smooth_colour = NULL,
  width = 800,
  height = 500,
  ticks = TRUE,
  ticklabels = "\%b \%d",
  date_constraint = FALSE,
  constraint_val = NULL,
  specified_type = NULL,
  a = NULL,
  b = NULL,
  data
)
}
\arguments{
\item{xaxis}{A character string that identifies the x axis variable (Date variable) to plot.}

\item{yaxis}{A list of lists that contains the entities to be used to plot the y axis. Each nested list is added as trace to the plot.
\itemize{
  \item y_column - The name of the variable to plot on yaxis.
  \item type - A character string that identifies the type of trace to add whether it is smoothed data, observed data or signal data. Based on that the template for plotting gets selected. Accepted values: "avg_data", "observed_data", "var_data", "doubling_time", "signal_data".
  \item name - A character string for the legend of the yaxis trace.
  \item color - A character string that indicates the color to use for the trace plot in the nested list in yaxis list that is to be added.
  \item width - A numeric value indicating the width of the trace plot in the nested list inside yaxis list.
  \item short_name - A character string indicating the type of data plotted in current trace. Accepted values: "projections", "Daily signal".
}}

\item{yaxis2}{Another list of lists to add y axis plots. The same list components as above}

\item{yaxis_button}{A boolean value that specifies whether to add button for yaxis trace}

\item{yaxis2_button}{A boolean value that specifies whether to add button for 2nd yaxis2 trace.}

\item{y_button_name}{A character string specifying the name for button}

\item{y2_button_name}{A character string specifying the name for button on second y axis}

\item{error_bands}{A boolean value that identifies whether we add confidence interval bands or not.}

\item{error_data}{A list that contains the confidence interval variable from the dataset for calculating the band width.}

\item{error_pct}{A boolean value representing whether the confidence intervals are calculated as percentages}

\item{error_col}{A hex value representing the color of the confidence bands}

\item{error_bars}{A boolean value that indicates whether error bars are to be formed on the yaxis trace.}

\item{level}{A boolean value that indicates whether to add a trace plot for level of detect.}

\item{level_upper}{A variable in the dataset which contains the upper limit value for level of detect.}

\item{level_lower}{A variable in the dataset which contains the lower limit value for level of detect.}

\item{level_col}{A hex value that indicates the color for the level of detect plot}

\item{titles}{A list that contains titles for xaxis, yaxis and the title of the plot.
\itemize{
  \item y - A character string for the label for the yaxis
  \item x - A character string for the label for the xaxis.
  \item title - A character string for the title of the plot.
}}

\item{vline}{A boolean value indicating whether there should be a vertical line added to the plot on the xaxis.}

\item{vline_date}{A date value that represents where on the xaxis the vertical line should start.}

\item{smooth}{A boolean value indicating whether to add a smoothed curve to the plot.}

\item{smooth_y}{A variable in the dataset to be used for creating the smooth curve.}

\item{smooth_bandwidth}{A numeric value indicating the bandwidth to use for smoothed curve.}

\item{smooth_name}{A character string representing the legend name for the smoothed curve.}

\item{smooth_colour}{A hex value representing the color of the smoothed curve.}

\item{width}{A numeric value representing the width of the plot.}

\item{height}{A numeric value representing the height of the plot.}

\item{ticks}{A boolean value indicating whether ticks should be visible on the axis.}

\item{ticklabels}{A character string indicates how to represent the tick labels for datetime values on xaxis. Ex: as month and year use: "%b %Y".}

\item{date_constraint}{A boolean value indicates whether the xaxis dates should be constrained on the default visible plot.}

\item{constraint_val}{A numeric value to subtract from the last date of input dataset to determine the start date in the input dataset.}

\item{specified_type}{A character string which contains the value of `type` entity present inside the yaxis list (another argument in this function) for the trace on which the error bars are required. Ex: "avg_data" is the value for `type` entity inside one of yaxis list which represents the trace plot for the average/ smoothed data.}

\item{a}{A numeric value representing the lower limit of yaxis scale}

\item{b}{A numeric value representing the upper limit of yaxis scale.}

\item{data}{A dataframe to be used as the source of input data for plotting.}
}
\value{
A plotly plot showing observed values as bar and smoothed estimates as line plot along with projections as line plot.
}
\description{
Accepts incoming data and creates plotly plots to show trend in waste water data and compare observed and predicted plots.
}
