% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_forecast.R
\name{short_term_forecast}
\alias{short_term_forecast}
\title{Accepts incoming data, cleans it and run EpiNow projections on it.}
\usage{
short_term_forecast(
  data,
  input,
  start_date,
  end_date,
  generation_time,
  incubation_period,
  reporting_delay,
  omit_last_date = FALSE,
  input_multiplier = 1,
  horizon = 14,
  rw = 7,
  gp = NULL,
  output = "projections",
  CrI = c(0.2, 0.5, 0.75, 0.9)
)
}
\arguments{
\item{data}{A data frame.}

\item{input}{A data frame column to run projections on.}

\item{start_date}{A start date for training data for EpiNow projections}

\item{end_date}{And end date for training data for EpiNow projections.}

\item{generation_time}{Time it takes on average for 1 infected person to cause secondary infections.}

\item{incubation_period}{Incubation period of the virus.}

\item{reporting_delay}{The delay in reporting an infection.}

\item{omit_last_date}{A boolean value specifying Whether to omit last date when running projections.}

\item{input_multiplier}{An initial number to multiply the input column with before running projections.}

\item{horizon}{The number of days to forecast into future.}

\item{rw}{Numeric step size of random walk (7 for weekly)}

\item{gp}{A list that defines the Gaussian process.}

\item{output}{Type of output to produce. There are 3 types: projection data frame, list of summary estimates, or both. Accepted values: "projections", "estimates", "both".}

\item{CrI}{A vector specifying which confidence interval to include in projections.}
}
\value{
The forecast data frame or list of summary estimates or Both
}
\description{
Accepts incoming data, cleans it and run EpiNow projections on it.
}
