% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_term_plot.R
\name{short_term_plot}
\alias{short_term_plot}
\title{Accepts incoming data and creates plotly plots to show observed as data as bar plots and smoothed estimates as line plot along with future projections.}
\usage{
short_term_plot(
  interval_num = 40,
  projections,
  levels = c("historic", "forecast"),
  obs_data,
  obs_column,
  forecast_type,
  start_date = first(as.Date(projections$date)),
  ylab,
  title,
  scale = FALSE,
  tick_period = "1 week",
  tick_labels_date = "\%b \%d",
  x_shift = 165,
  y_shift = 0,
  annotation_text = "*Shaded area represents the 75\% credible region",
  CrI = 75,
  CrIp = 25
)
}
\arguments{
\item{interval_num}{A numeric value denoting how many dates to show on x axis prior to the last date in data.}

\item{projections}{A data frame that contains EpiNow projections.}

\item{levels}{A vector that is used to create legend for historic data and projected data.}

\item{obs_data}{A data frame that consists of observed historic data.}

\item{obs_column}{A character string that specifies which column to select for plotting from observed data.}

\item{forecast_type}{A character string specifying what type of EpiNow projections to plot. Acceptable values are reported_cases, R, infection, growth_rate.}

\item{start_date}{A date value specifying what date to include as start date on x axis of the plot.}

\item{ylab}{A character string specifying the label for y axis.}

\item{title}{A character string specifying the title of the plot.}

\item{scale}{A boolean value specifying whether to scale the plot}

\item{tick_period}{A character string specifying the intervals between two consecutive dates on x axis}

\item{tick_labels_date}{A character string specifying the tick labels for dates on x axis.}

\item{x_shift}{A numeric value to show where to shift the annotation text on the x axis}

\item{y_shift}{A numeric value to show where to shift the annotation text on the y axis}

\item{annotation_text}{A character string that provides annotation text to show on the plot for description.}

\item{CrI}{A numeric value specifying which confidence interval to include.}

\item{CrIp}{A numeric value used for denoting the lower level of confidence interval desired.}
}
\value{
A plotly plot showing observed values as bar and smoothed estimates as line plot along with projections as line plot.
}
\description{
Accepts incoming data and creates plotly plots to show observed as data as bar plots and smoothed estimates as line plot along with future projections.
}
