% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vline.R
\name{vline}
\alias{vline}
\title{Accepts an x asxis value and a string representing color value to return a list that can be used to add as a vertical line trace to the plotly plot.}
\usage{
vline(x = 0, color = "blue")
}
\arguments{
\item{x}{A value on the xaxis where the vertical line should start.}

\item{color}{A string value representing the color of the vertical line to be added}
}
\value{
A list that is used by plotly to add as trace to a plotly figure to add a vertical line.
}
\description{
Accepts an x asxis value and a string representing color value to return a list that can be used to add as a vertical line trace to the plotly plot.
}
