% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastewater_prep.R
\name{wastewater_prep}
\alias{wastewater_prep}
\title{Accepts incoming wastewater data and cleans it.}
\usage{
wastewater_prep(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
The final cleaned data frame which contains renamed variables, formatted date, new variables for rolling averages for 5, 7, and 10 days and daily rate of change in signal.
}
\description{
Accepts incoming wastewater data and cleans it.
}
\examples{
wastewater_prep(read.csv("https://raw.githubusercontent.com/Big-Life-Lab/PHESD/main/Wastewater/Ottawa/Data/wastewater_virus.csv"))
}
